Don't upload if the object already exists and has the right md5hash.

+++ src/main/scala/ec2/s3cache.scala
@@ -23,12 +23,18 @@ object S3Cache extends AWSConnection {
 			md5Cache.get(fileMd5) match {
 				case Some(url) => url
 				case None => {
-					logger.info("Uploading " + file + " to S3")
-					val obj = new S3Object(file)
-					obj.setKey(fileMd5)
-					obj.setAcl(AccessControlList.REST_CANNED_PUBLIC_READ)
+          val existingObject: Option[S3Object] = try Some(s3Service.getObjectDetails(bucket, fileMd5)) catch {
+            case _: org.jets3t.service.S3ServiceException => None
+          }
 
-					s3Service.putObject(bucket, obj)
+          if(! existingObject.map(_.getMd5HashAsHex == fileMd5).getOrElse(false)) {
+            logger.info("Uploading " + file + " to S3")
+            val obj = new S3Object(file)
+            obj.setKey(fileMd5)
+            obj.setAcl(AccessControlList.REST_CANNED_PUBLIC_READ)
+
+            s3Service.putObject(bucket, obj)
+          }
 					val url =  "http://s3.amazonaws.com/" + bucketName + "/" + fileMd5
 					md5Cache.put(fileMd5, url)
 					url
